Finishes off misc.py
