+++ src/metaArray/__init__.py
@@ -3,7 +3,12 @@ __version__ = '2.0.0'
 
 
 # Import required functionality
-from metaArray.core import metaArray
+from .core import metaArray
 
 # Examples
-from metaArray.example import demo
+from .example import demo
+
+# Helper functions from misc
+from .misc import dirPath, filePath, file_list, buffered_search
+from .misc import gettypecode, unit_prefix, resample
+from .misc import cplx_trig_func, mother_morlet, spline_resize
